removed onednn_inductor_quantizer insted use x86_inductor_quantizer
